requirements-dev.txt
fastapi[all]
passlib[bcrypt]
pre-commit
psycopg2
pydantic-settings
pyjwt
sqlalchemy
